requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.4.0
pymongo==4.6.1
twilio==8.10.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
python-dateutil==2.8.2
phonenumbers==8.13.25
pydantic==2.5.0
groq==0.22.0
